Fully functioning damped solve routines to tackle buckling.
Adaptive timestepping solve method which utilises damped versions
of the FvK equations in order to stabilise the local buckling that
occurs with swelling.

Additionally, there is new steady/unsteady output functionality in
the rectangular driver code.

A circular driver code has been included to test the damped
equations in a simple deflection case using normal and curved
elements.

Removed remnants of Davids driver code as it doesn't belong here.

=====================================================================

On branch master
Your branch is up-to-date with 'origin/master'.

Changes to be committed:
	modified:   Makefile.am
	modified:   Makefile.in
	deleted:    README.md
	deleted:    circular_disc_free_foeppl_von_karman.cc
	deleted:    circular_disc_free_foeppl_von_karman_simon.cc
	new file:   circular_disc_fvk.cc
	deleted:    circular_disc_uniform_pressure_foeppl_von_karman.cc
	deleted:    circular_disc_uniform_pressure_foeppl_von_karman_demo.cc
	new file:   plot.bash
	modified:   rectangular_plate_foeppl_von_karman.cc
	new file:   rectangular_run.bash